moved template files to template folder
